Added laser cut pattern DXF file for v2.6 design
This file is a single DXF containing laser cut patterns for all parts
needed for v2.6 (parts are to be cut from a variety of plastic types and
thicknesses).   Looking at groups from top to bottom, the first group
(the shell) is 3mm-thick translucent-white polypropylene.  Group 2 is
3mm clear cast acrylic.  Group three is 3mm white acrylic.  Group 4 is
1.5mm clear acrylic.  Group 6 is 1.5mm white acrylic.   Group 7 is 6mm
thick clear acrylic.